pyproject.toml
[project]
name = "invoice-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.2",
]
